Add LIS classes

+++ source/eis/rechunk.py
@@ -42,14 +42,15 @@ class Rechunker:
 
     def rechunk( self, chunk_sizes: Dict[str,int], **kwargs ):
         t0 = time.time()
-        max_memory =   kwargs.pop( 'max_memory', "500MB" )
+        max_memory =   kwargs.pop( 'max_memory', 100 ) * 1000 * 1000
         target_store = kwargs.pop( 'target_store', self.data_dir )
         temp_store =   kwargs.pop( 'temp_store', self.cache_dir )
         chunks = self.get_chunks( chunk_sizes )
         if isinstance( target_store, str ):
             target_store = f"{target_store}/{self.name}.zarr"
-            shutil.rmtree( target_store, ignore_errors= True )
-            print( f"Writing result to {target_store} with max-memory-per-worker set to {max_memory}" )
+            if target_store.startswith("/"):
+                shutil.rmtree( target_store, ignore_errors= True )
+            print( f"Writing result to {target_store} with max-memory-per-worker set to {max_memory} bytes" )
         if isinstance( temp_store, str):
             temp_store =  f"{temp_store}/{self.name}.zarr"
             shutil.rmtree( temp_store, ignore_errors= True )
